+++ CHANGELOG.md
@@ -6,8 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Next
+
+## Added
 - Add the ability to smooth raw signals using convolution with a Gaussian or flat kernel. 
 - Add the ability to smooth raw signals using a Savitzky-Golay filter
+- Ability to calculate the region-level spearman correlations for a set of samples in `bwtools_query`
 
 ## 0.2.4 - 2021-10-13