+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add the ability to smooth raw signals using convolution with a Gaussian or flat kernel. 
 - Add the ability to smooth raw signals using a Savitzky-Golay filter
 - Ability to calculate the region-level spearman correlations for a set of samples in `bwtools_query`
+- Add calculations of the traveling ratio (10.1016/j.molcel.2008.12.021)
+- Add discovery of a local max within a region
 
 ### Changed
 - Output of `bwtools_query` is now compressed (gzip). File extension changed to `.tsv.gz`